requirements.txt
asgiref
Django~=5.0.6
django-cors-headers
djangorestframework~=3.15.1
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv~=1.0.1
google
geopy~=2.4.1